requirements.txt
# Confluent Kafka
confluent-kafka==2.3.0
confluent-kafka[avro]==2.3.0

# Google Cloud
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.13.0
# Note: BigQuery ML functionality is included in google-cloud-bigquery
google-cloud-storage==2.14.0

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0

# 数据处理
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# 机器学习
# TensorFlow 2.16+ supports Python 3.11
tensorflow>=2.16.0
keras>=2.16.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1

# 地理和地图
geopy==2.4.1
shapely==2.0.2

# 时间处理
pytz==2023.3

